Always request a new transfer in midih_xfer_cb
If the midih_xfer_cb function is called, it means
the last transfer completed, even if the number
of bytes transferred was 0. I saw in my testing that
the RP2040 HW does not complete the transfer for a NAK
response, but it seems that it can complete it with 0
bytes transferred per usb_midi_host Issue #2. This
change should fix it.